requirements.txt
flake8
pytest
pyyaml
yattag
